Make parts of Fusion Index available to Spatial for dynamic fusion
